requirements.txt
libusb_package==1.0.26.0
monitorcontrol==3.0.1
pyusb==1.2.1
